Created Grupos.SensorsList, Grupos, __old, Grupos.GroupTemplate, DashBoard.SensorTemplate, __old.DashBoard_V2, Grupos.SensorsList.SensorTemplate, __old.DashBoard_V2.ItemTemplate1 and __old.DashBoard; Edited settings, Login, Sensores.SensorTemplate, DashBoard, MenuLateral and routes; Deleted DashBoard_V2, DashBoard_V3.SensorTemplate, DashBoard_V2.ItemTemplate1 and DashBoard_V3
