+++ .anvil_editor.yaml
@@ -1,21 +1,26 @@
 unique_ids:
   forms:
+    Grupos.SensorsList: 6R3ESLOZ6PJZHFEJQA4AJU7L2GFGQ5RO
+    Grupos: D5ZU3U7LYEJCNTS5US6EWKLAXMYRPOEH
     Usuarios.EditGroups: QYWXE2JEQMF2D2HB5RALXMZ6V33ZMC7I
     Usuarios.CreateUser: A65IGXZDR5DSCOFWAVZGQY6FD5SOQWII
+    __old: K4FT2R65H32NNPCG44VMS4Z3QFL3CJF4
     Usuarios.UserTemplate: VVOYFK5KGNHLEPPKSHK2XXQV7663VXV3
     Login: USBYLWDVV3EOLILIJAMPBY3DL5MTA2SX
+    Grupos.GroupTemplate: 3MAA7M5GBN4KGYCPBS5KCZMTL5OAIUGL
+    DashBoard.SensorTemplate: H2W62VRZCAN534TIAEU6K3L6IQOPUACZ
+    __old.DashBoard_V2: JXU6QQVYVPBJFNCCQOJ4CDYV5IEN36BV
     Usuarios: KYFM6KR7CZSBQN54EMAYHIAGETR2A7ZX
-    DashBoard_V2: JXU6QQVYVPBJFNCCQOJ4CDYV5IEN36BV
     Sensores: E6KCOK5IZGTVCMVFVMUC6ZAJ5DYRPWNT
-    DashBoard_V3.SensorTemplate: H2W62VRZCAN534TIAEU6K3L6IQOPUACZ
+    Grupos.SensorsList.SensorTemplate: EBO3MNCSKDGOUNEYE2ZIEALH4CID3RY6
     Sensores.TypeTemplate: 7TJ24RY23T3OPTRTWEDYYXRBSKX4YPHB
     Usuarios.EditGroups.GroupTemplate: TUPMA4PYU4Q53AFCJMUHDNTIJMX2HURT
-    DashBoard_V2.ItemTemplate1: O6RDPWEWU44GOOMLGYWTHXXEVSEMNVQK
+    __old.DashBoard_V2.ItemTemplate1: O6RDPWEWU44GOOMLGYWTHXXEVSEMNVQK
     Sensores.SensorTemplate: T7GEIYBW2O57AYLS4AMQ4CBFYDMYTKD3
     Sensores.EditarSensor: G7SWHJH3XD4CJYCU7S6DMF6LQN7MOWPI
-    DashBoard: Y3KBOBKYSGDU7KXYO3SLMRCPMDDWKC5X
+    DashBoard: CRGNCOVOHEQ3DY2KUS4ZAMKGDQIH2BSE
+    __old.DashBoard: Y3KBOBKYSGDU7KXYO3SLMRCPMDDWKC5X
     MenuLateral: MVASCYYRSIMLJJJRJXBM4BVO4VAQ2TMH
-    DashBoard_V3: CRGNCOVOHEQ3DY2KUS4ZAMKGDQIH2BSE
   modules:
     functions: TMINS64CGVKJWRGQRWB2UP7YK5OYPQRA
     routes: LTWIRSS4WZ3VOBXGCMJ5MOS7WYBH5LTZ
