+++ anvil.yaml
@@ -14,6 +14,7 @@ name: MC855
 runtime_options: {version: 2, server_version: python3-sandbox, client_version: '3',
   server_spec: null}
 metadata: {}
+startup_form: Login
 native_deps: {head_html: ''}
 db_schema:
   revenue:
