+++ client_code/DashBoard/SensorTemplate.yaml
@@ -1,7 +1,7 @@
 components:
 - type: ColumnPanel
   properties: {role: card, tooltip: '', border: '', foreground: '', visible: true,
-    wrap_on: mobile, col_spacing: medium, spacing_above: small, col_widths: '{"KIUFGS":20,"KZXVAT":19,"OLPLEK":10,"NQCZFT":40,"BEKICA":10}',
+    wrap_on: mobile, col_spacing: medium, spacing_above: small, col_widths: '{"KIUFGS":10,"KZXVAT":20,"DQQCZM":30}',
     spacing_below: small, background: ''}
   name: card_4
   layout_properties:
@@ -10,10 +10,6 @@ components:
     full_width_row: true
   components:
   - type: Label
-    properties: {}
-    name: label_1
-    layout_properties: {grid_position: 'TCEHVV,OLPLEK'}
-  - type: Label
     properties: {role: null, align: left, tooltip: '', border: '', foreground: '',
       visible: true, text: Sensor, font_size: null, font: '', spacing_above: small,
       icon_align: left, spacing_below: small, italic: false, background: '', bold: true,
@@ -21,7 +17,7 @@ components:
     name: label_name
     layout_properties:
       col_widths: {}
-      grid_position: TCEHVV,NQCZFT FMVBIR,KIUFGS XJRHHG,YFUACB
+      grid_position: FMVBIR,KIUFGS XJRHHG,YFUACB
   - type: Label
     properties: {role: null, align: left, tooltip: '', border: '', foreground: '',
       visible: true, text: Medida, font_size: null, font: '', spacing_above: small,
@@ -30,14 +26,14 @@ components:
     name: label_name_copy
     layout_properties:
       col_widths: {}
-      grid_position: TCEHVV,NQCZFT FMVBIR,KIUFGS WVOENH,IDWXIG
+      grid_position: FMVBIR,KIUFGS WVOENH,IDWXIG
   - type: Label
     properties: {role: null, align: left, tooltip: '', border: '', foreground: '',
       visible: true, text: Limite Inferior, font_size: null, font: '', spacing_above: small,
       icon_align: left, spacing_below: small, italic: false, background: '', bold: true,
       underline: false, icon: ''}
     name: label_current_temp
-    layout_properties: {grid_position: 'TCEHVV,NQCZFT FMVBIR,KIUFGS TRARZJ,YKALXU'}
+    layout_properties: {grid_position: 'FMVBIR,KIUFGS TRARZJ,YKALXU'}
   - type: Label
     properties: {role: null, align: left, tooltip: '', border: '', foreground: '',
       visible: true, text: Limite Superior, font_size: null, font: '', spacing_above: small,
@@ -46,62 +42,51 @@ components:
     name: label_limit
     layout_properties:
       col_widths: {}
-      grid_position: TCEHVV,NQCZFT FMVBIR,KIUFGS GPLNAR,MYWHHT
+      grid_position: FMVBIR,KIUFGS GPLNAR,MYWHHT
   - type: Label
     properties: {role: null, align: left, tooltip: '', border: '', foreground: '',
       visible: true, text: Ultimo Valor, font_size: null, font: '', spacing_above: small,
       icon_align: left, spacing_below: small, italic: false, background: '', bold: true,
       underline: false, icon: ''}
     name: label_max_temp
-    layout_properties: {grid_position: 'TCEHVV,NQCZFT FMVBIR,KIUFGS FENQAV,OIZTJZ'}
+    layout_properties: {grid_position: 'FMVBIR,KIUFGS FENQAV,OIZTJZ'}
   - type: Label
     properties: {}
     name: name
-    layout_properties: {grid_position: 'TCEHVV,NQCZFT FMVBIR,KZXVAT GHKUMC,SWVUHT'}
+    layout_properties: {grid_position: 'FMVBIR,KZXVAT GHKUMC,SWVUHT'}
     data_bindings: []
   - type: Label
     properties: {}
     name: type
     layout_properties:
       col_widths: {}
-      grid_position: TCEHVV,NQCZFT FMVBIR,KZXVAT DNQBXQ,YCZMJZ
+      grid_position: FMVBIR,KZXVAT DNQBXQ,YCZMJZ
     data_bindings: []
   - type: Label
     properties: {}
     name: bottom_limit
     layout_properties:
       col_widths: {}
-      grid_position: TCEHVV,NQCZFT FMVBIR,KZXVAT LAWTLX,HIWWEN
+      grid_position: FMVBIR,KZXVAT LAWTLX,HIWWEN
     data_bindings: []
   - type: Label
     properties: {}
     name: upper_limit
     layout_properties:
       col_widths: {}
-      grid_position: TCEHVV,NQCZFT FMVBIR,KZXVAT AUUSFK,PHTAVV
+      grid_position: FMVBIR,KZXVAT AUUSFK,PHTAVV
     data_bindings: []
   - type: Label
     properties: {}
     name: last_value
-    layout_properties: {grid_position: 'TCEHVV,NQCZFT FMVBIR,KZXVAT QPXRLN,FGFJVC'}
+    layout_properties: {grid_position: 'FMVBIR,KZXVAT QPXRLN,FGFJVC'}
     data_bindings: []
-  - type: Label
-    properties: {}
-    name: label_2
-    layout_properties: {grid_position: 'TCEHVV,BEKICA', full_width_row: true}
-  - type: ColumnPanel
-    properties: {role: card, tooltip: '', border: '', foreground: '', visible: true,
-      wrap_on: mobile, col_spacing: medium, spacing_above: small, col_widths: '{}',
-      spacing_below: small, background: ''}
-    name: column_panel_1
-    layout_properties: {grid_position: 'MQYOPW,JSGSUW'}
-    components:
-    - type: Plot
-      properties: {height: 216.33333206176758}
-      name: plot_1
-      layout_properties: {grid_position: 'AVVSTP,MRSNSY', full_width_row: true}
-      event_bindings: {}
-      data_bindings: []
+  - type: Plot
+    properties: {height: 216.33333206176758}
+    name: plot_1
+    layout_properties: {grid_position: 'FMVBIR,DQQCZM', full_width_row: true}
+    event_bindings: {}
+    data_bindings: []
 container:
   type: ColumnPanel
   properties: {col_widths: '{}'}
