+++ client_code/DashBoard/form_template.yaml
@@ -1,94 +1,31 @@
-container:
-  type: HtmlTemplate
-  properties: {html: '@theme:standard-page.html'}
 components:
 - type: ColumnPanel
-  properties: {col_widths: '{"BNRDLT":30,"ZDTODK":30}'}
+  properties: {col_widths: '{}'}
   name: content_panel
   layout_properties: {slot: default}
   components:
-  - type: ColumnPanel
-    properties: {role: card, col_widths: '{}'}
-    name: card_6
-    layout_properties: {grid_position: 'ITUJQM,BNRDLT', full_width_row: true}
-    components:
-    - type: Label
-      properties: {role: text, align: left, tooltip: '', border: '', foreground: '',
-        visible: true, text: 'Temperatura Atual
-
-          ', font_size: null, font: '', spacing_above: small, icon_align: left, spacing_below: none,
-        italic: false, background: '', bold: true, underline: false, icon: ''}
-      name: label_3
-      layout_properties: {grid_position: 'APMUTY,TWEYQS'}
-      data_bindings: []
-      event_bindings: {}
-    - type: Label
-      properties: {role: null, align: left, tooltip: '', border: '', foreground: '',
-        visible: true, text: '', font_size: null, font: '', spacing_above: none, icon_align: left,
-        spacing_below: small, italic: false, background: '', bold: false, underline: false,
-        icon: ''}
-      name: last_temp
-      layout_properties: {grid_position: 'KWMDOQ,JIOVNV'}
-      event_bindings: {}
-  - type: ColumnPanel
-    properties: {role: card, col_widths: '{}'}
-    name: card_7
-    layout_properties: {grid_position: 'ITUJQM,ZDTODK', full_width_row: true}
-    components:
-    - type: Label
-      properties: {role: text, align: left, tooltip: '', border: '', foreground: '',
-        visible: true, text: 'Maior temperatura do Dia
-
-          ', font_size: null, font: '', spacing_above: small, icon_align: left, spacing_below: none,
-        italic: false, background: '', bold: true, underline: false, icon: ''}
-      name: label_4
-      layout_properties: {grid_position: 'PKAGQX,FYTMDT'}
-    - type: Label
-      properties: {role: null, align: left, tooltip: '', border: '', foreground: '',
-        visible: true, text: '', font_size: null, font: '', spacing_above: none, icon_align: left,
-        spacing_below: small, italic: false, background: '', bold: false, underline: false,
-        icon: ''}
-      name: max_daily_temp
-      layout_properties: {grid_position: 'TMMHLS,FVFKVQ'}
-  - type: ColumnPanel
-    properties: {role: card, tooltip: '', border: '', foreground: '', visible: true,
-      wrap_on: mobile, col_spacing: medium, spacing_above: small, col_widths: '{}',
-      spacing_below: small, background: ''}
-    name: card_4
-    layout_properties: {grid_position: 'HPJDEK,XCBAHY', full_width_row: true}
-    components:
-    - type: Plot
-      properties: {height: 270}
-      name: plot_1
-      layout_properties: {grid_position: 'LFQCZD,DVQKEA'}
-      event_bindings: {}
-  - type: ColumnPanel
-    properties: {role: card, col_widths: '{}'}
-    name: card_2
-    layout_properties: {grid_position: 'UJVGGL,YVUGLI', full_width_row: true}
-    components:
-    - type: Plot
-      properties: {height: 274}
-      name: plot_2
-      layout_properties: {grid_position: 'BGZEUQ,EGLLGT'}
+  - type: RepeatingPanel
+    properties: {role: null, tooltip: '', border: '', foreground: '', items: null,
+      visible: true, spacing_above: small, spacing_below: small, item_template: DashBoard.SensorTemplate,
+      background: ''}
+    name: repeating_panel_1
+    layout_properties: {grid_position: 'QUSQBZ,XTGDBB'}
 - type: FlowPanel
   properties: {}
   name: navbar_links
   layout_properties: {slot: nav-right}
 - type: Label
   properties: {role: null, align: left, tooltip: '', border: '', foreground: '', visible: true,
-    text: 'Painel de Temperaturas
-
-      ', font_size: null, font: '', spacing_above: small, icon_align: left, spacing_below: small,
-    italic: false, background: '', bold: false, underline: false, icon: ''}
+    text: Dashboard, font_size: null, font: '', spacing_above: small, icon_align: left,
+    spacing_below: small, italic: false, background: '', bold: false, underline: false,
+    icon: ''}
   name: label_1
   layout_properties: {slot: title}
 - type: form:MenuLateral
   properties: {}
   name: menu_lateral_1
   layout_properties: {slot: left-nav}
-- type: Timer
-  properties: {interval: 1}
-  name: timer_1
-  event_bindings: {tick: timer_1_tick}
+container:
+  type: HtmlTemplate
+  properties: {html: '@theme:standard-page.html'}
 is_package: true
