--- client_code/DashBoard_V3/form_template.yaml
@@ -1,31 +1,0 @@
-components:
-- type: ColumnPanel
-  properties: {col_widths: '{}'}
-  name: content_panel
-  layout_properties: {slot: default}
-  components:
-  - type: RepeatingPanel
-    properties: {role: null, tooltip: '', border: '', foreground: '', items: null,
-      visible: true, spacing_above: small, spacing_below: small, item_template: DashBoard_V3.SensorTemplate,
-      background: ''}
-    name: repeating_panel_1
-    layout_properties: {grid_position: 'QUSQBZ,XTGDBB'}
-- type: FlowPanel
-  properties: {}
-  name: navbar_links
-  layout_properties: {slot: nav-right}
-- type: Label
-  properties: {role: null, align: left, tooltip: '', border: '', foreground: '', visible: true,
-    text: Dashboard, font_size: null, font: '', spacing_above: small, icon_align: left,
-    spacing_below: small, italic: false, background: '', bold: false, underline: false,
-    icon: ''}
-  name: label_1
-  layout_properties: {slot: title}
-- type: form:MenuLateral
-  properties: {}
-  name: menu_lateral_1
-  layout_properties: {slot: left-nav}
-container:
-  type: HtmlTemplate
-  properties: {html: '@theme:standard-page.html'}
-is_package: true
