+++ client_code/Grupos/GroupTemplate/form_template.yaml
@@ -1,0 +1,28 @@
+container: {type: DataRowPanel}
+components:
+- type: Button
+  properties: {role: null, align: center, tooltip: '', border: '', enabled: true,
+    foreground: '', visible: true, text: Deletar, font_size: null, font: '', spacing_above: small,
+    icon_align: left, spacing_below: small, italic: false, background: '', bold: false,
+    underline: false, icon: ''}
+  name: delete_button
+  layout_properties: {column: LRMNKL}
+  event_bindings: {click: delete_button_click}
+- type: Button
+  properties: {role: null, align: center, tooltip: '', border: '', enabled: true,
+    foreground: '', visible: true, text: Sensores, font_size: null, font: '', spacing_above: small,
+    icon_align: left, spacing_below: small, italic: false, background: '', bold: false,
+    underline: false, icon: ''}
+  name: sensors_button
+  layout_properties: {column: WYHREP}
+  event_bindings: {click: sensors_button_click}
+- type: Label
+  properties: {role: null, align: center, tooltip: '', border: '', foreground: '',
+    visible: true, text: '', font_size: null, font: '', spacing_above: small, icon_align: left,
+    spacing_below: small, italic: false, background: '', bold: false, underline: false,
+    icon: ''}
+  name: group_name
+  layout_properties: {column: VNBLUM}
+  data_bindings:
+  - {property: text, code: 'self.item[''group_name'']'}
+is_package: true
