+++ client_code/Grupos/SensorsList/SensorTemplate/form_template.yaml
@@ -1,0 +1,24 @@
+container:
+  type: ColumnPanel
+  properties: {col_widths: '{}'}
+components:
+- type: Label
+  properties: {role: null, align: center, tooltip: '', border: '', foreground: '',
+    visible: true, text: '', font_size: null, font: '', spacing_above: small, icon_align: left,
+    spacing_below: small, italic: false, background: '', bold: false, underline: false,
+    icon: ''}
+  name: sensor_name
+  layout_properties: {grid_position: 'JTZUCH,GUEZJC'}
+  data_bindings:
+  - {property: text, code: 'self.item[''sensor_name'']'}
+- type: CheckBox
+  properties: {role: null, align: center, tooltip: '', border: '', enabled: true,
+    foreground: '', allow_indeterminate: false, visible: true, text: '', font_size: null,
+    font: '', spacing_above: small, spacing_below: small, italic: false, background: '',
+    bold: false, checked: false, underline: false}
+  name: check_box_1
+  layout_properties: {grid_position: 'JTZUCH,YTLLAX'}
+  data_bindings:
+  - {property: checked, code: 'self.item[''in_group'']', writeback: true}
+is_package: true
+item_type: {table_id: null}
