+++ client_code/Grupos/SensorsList/form_template.yaml
@@ -1,0 +1,9 @@
+container:
+  type: ColumnPanel
+  properties: {col_widths: '{}'}
+components:
+- type: RepeatingPanel
+  properties: {item_template: Grupos.SensorsList.SensorTemplate}
+  name: repeating_panel_1
+  layout_properties: {grid_position: 'XLDNOW,YPUMHX'}
+is_package: true
