+++ client_code/Login/__init__.py
@@ -28,14 +28,14 @@ class Login(LoginTemplate):
       self.message_label.text = "Usuário/Senha incorreto"
     else:
       localStorage.setItem('access_token', token["access_token"])
-      open_form('Sensores')
+      open_form('DashBoard')
             
   def verify_session(self, **event_args):
       valid_session = functions.validate_session()
       
       if(valid_session != False):
         print("User already logged in")
-        open_form('DashBoard_V2')
+        open_form('DashBoard')
