+++ client_code/MenuLateral/form_template.yaml
@@ -11,6 +11,14 @@ components:
   name: button_2
   layout_properties: {grid_position: 'CQLUKC,SNTEEY'}
   event_bindings: {click: button_2_click}
+- type: Button
+  properties: {role: null, align: center, tooltip: '', border: '', enabled: true,
+    foreground: '', visible: true, text: Grupos, font_size: null, font: '', spacing_above: small,
+    icon_align: left, spacing_below: small, italic: false, background: '', bold: false,
+    underline: false, icon: ''}
+  name: button_4
+  layout_properties: {grid_position: 'CKEGEW,UXMENN'}
+  event_bindings: {click: button_4_click}
 - type: ColumnPanel
   properties: {col_widths: '{}'}
   name: column_panel_1
