+++ client_code/Sensores/SensorTemplate/form_template.yaml
@@ -80,11 +80,12 @@ components:
     event_bindings: {click: edit_button_click}
   - type: Button
     properties: {role: null, align: center, tooltip: '', border: '', enabled: true,
-      foreground: '', visible: true, text: DashBoards, font_size: null, font: '',
-      spacing_above: small, icon_align: left, spacing_below: small, italic: false,
-      background: '', bold: false, underline: false, icon: ''}
-    name: button_2
+      foreground: '', visible: true, text: DashBoard, font_size: null, font: '', spacing_above: small,
+      icon_align: left, spacing_below: small, italic: false, background: '', bold: false,
+      underline: false, icon: ''}
+    name: dashboard_button
     layout_properties: {grid_position: 'GIONTF,FSMRXB FTVWEX,RWWPEY'}
+    event_bindings: {click: dashboard_button_click}
   - type: Label
     properties: {}
     name: space_2
