+++ client_code/__old/DashBoard_V2/form_template.yaml
@@ -9,7 +9,7 @@ components:
   components:
   - type: RepeatingPanel
     properties: {role: null, tooltip: '', border: '', foreground: '', items: null,
-      visible: true, spacing_above: small, spacing_below: small, item_template: DashBoard_V2.ItemTemplate1,
+      visible: true, spacing_above: small, spacing_below: small, item_template: __old.DashBoard_V2.ItemTemplate1,
       background: ''}
     name: repeating_panel_1
     layout_properties: {grid_position: 'QUSQBZ,XTGDBB'}
