+++ client_code/__old/__init__.py
@@ -1,0 +1,13 @@
+from ._anvil_designer import __oldTemplate
+from anvil import *
+import anvil.tables as tables
+import anvil.tables.query as q
+from anvil.tables import app_tables
+
+class __old(__oldTemplate):
+  def __init__(self, **properties):
+    # Set Form properties and Data Bindings.
+    self.init_components(**properties)
+
+    # Any code you write here will run when the form opens.
+
